feat(wm): add threshold-based layout_options_rules for workspaces
Adds layout_options_rules to workspace configuration, allowing layout_options to dynamically change based on container count. Uses the same threshold semantics as layout_rules: when container count >= threshold, the highest matching rule fully replaces the base layout_options.
